+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import common from './common';
+
+export default combineReducers({
+  common,
+});
